refactor(chat): extract chat_manager/reads.rs (#292)
Move the registry/queue reads, enriched registry reads, and in-memory
cache sync + broadcast methods off the facade impl into their own
submodule.
